Add files via upload
CS 300 Project Two. CPP file that will take a file name from input, read the file line by line, and load it into a BinarySearch Tree. User will be able to perform various tasks on the BST such as display a sorted list of all of the courses in the file to the user, allow the search to search for a specific course and display the course's information